+++ src/locales/langs/en-us.ts
@@ -231,7 +231,6 @@ const local: App.I18n.Schema = {
     'iframe-page': 'Iframe',
     home: 'Home',
     'base64-converter': 'Base64 Converter',
-    'character-generator': 'Character Generator (test)',
     'btc-course-flow': 'BTC Course Flow',
     'effects-generator': 'Effects Generator',
     'image-compressor': 'Image Compressor',